Fixed Breaking Change for MartenCommandException
