+++ src/Marten.Testing/Acceptance/computed_indexes.cs
@@ -237,7 +237,7 @@ namespace Marten.Testing.Acceptance
 
                 // Inserting the same original string should throw
                 session.Store(item2);
-                Assert.Throws<MartenCommandException>(() => session.SaveChanges()).Message.ShouldContain("duplicate");
+                Assert.Throws<Marten.Exceptions.MartenCommandException>(() => session.SaveChanges()).Message.ShouldContain("duplicate");
             }
         }
 
@@ -318,7 +318,7 @@ namespace Marten.Testing.Acceptance
                 // Inserting the same string but all uppercase should throw because
                 // the index is stored with lowcased value
                 session.Store(item);
-                Assert.Throws<MartenCommandException>(() => session.SaveChanges()).Message.ShouldContain("duplicate");
+                Assert.Throws<Marten.Exceptions.MartenCommandException>(() => session.SaveChanges()).Message.ShouldContain("duplicate");
             }
         }
