+++ src/Marten/MartenCommandException.cs
@@ -3,15 +3,54 @@ using Npgsql;
 
 namespace Marten
 {
-    [Obsolete("This class is Obsolete, please use Marten.Exceptions.MartenCommandException.")]
-    public class MartenCommandException: Exceptions.MartenCommandException
+    /// <summary>
+    /// Wraps the Postgres command exceptions. Unifies exception handling and brings additonal information.
+    /// </summary>
+    [Obsolete("This class is Obsolete, please use Marten.Exceptions.MartenCommandException. It might be removed in v4.0.")]
+    public class MartenCommandException: Exception
     {
-        public MartenCommandException(NpgsqlCommand command, Exception innerException) : base(command, innerException)
+        /// <summary>
+        /// Failed Postgres command
+        /// </summary>
+        public NpgsqlCommand Command { get; }
+
+        protected static string ToMessage(
+            NpgsqlCommand command,
+            string prefix = null
+        )
         {
+            if (prefix != null)
+            {
+                prefix = $"{prefix}${Environment.NewLine}";
+            }
+
+            return $"Marten Command Failure:${Environment.NewLine}{prefix}{command.CommandText}${Environment.NewLine}${Environment.NewLine}";
         }
 
-        public MartenCommandException(NpgsqlCommand command, Exception innerException, string prefix) : base(command, innerException, prefix)
+        /// <summary>
+        /// Creates MartenCommandException based on the command and innerException information with formatted message.
+        /// </summary>
+        /// <param name="command">failed Postgres command</param>
+        /// <param name="innerException">internal exception details</param>
+        public MartenCommandException(NpgsqlCommand command, Exception innerException)
+            : base(ToMessage(command) + innerException.Message, innerException)
         {
+            Command = command;
+        }
+
+        /// <summary>
+        /// Creates MartenCommandException based on the command and innerException information with formatted message.
+        /// </summary>
+        /// <param name="command">failed Postgres command</param>
+        /// <param name="innerException">internal exception details</param>
+        /// <param name="prefix">prefix that will be added to message</param>
+        public MartenCommandException(
+            NpgsqlCommand command,
+            Exception innerException,
+            string prefix
+        ) : base(ToMessage(command, prefix) + innerException.Message, innerException)
+        {
+            Command = command;
         }
     }
 }
